Functioning VS code
Needs to implement something with tabs, they're too cool to get rid of. There is a total function that could be displayed large or you could put the values in a table row for each country like a dictionary. I still need to create functionality for storing the data, my idea was to parse the timestamp to only the date, no time, and store the country names in a day/time stamp collection and each document is a country actually for example. we could then perhaps do a dropdown to display that day's data? 
//python code
ref_station_type = db.collection(u'PARSEDTIMESTAMP').document(u'' + self.country)

If a function is made for displaying the table for the current day with the total at the bottom, it could easily display the past ones depending on the dropdown selection. 
The last thing needed on the leaflet map, is simply to limit the pan bounds, once the map is instantiated, set the starting coords, bounds, and zoom to something that shows all points, then set the pan bounds to those.

Thanks for reading this,
xoxoxox